# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type User struct {
	UserID    pgtype.UUID
	FirstName string
	LastName  string
	Email     string
	Phone     string
	Age       int32
	Status    int32
}